Add files via upload
Cell maps can be displayed by SDL2 Windows, with black representing living cells and white representing dead cells. In the beginning, there will be a page where the game is loading. When all cells on the map are dead, there will be a screen indicating the end of the game.